fix: commits with --amend
