+++ lib/git-helpers.js
@@ -1,5 +1,6 @@
 'use strict'
 
+const url = require('url')
 const exec = require('child_process').execSync
 
 const _ = require('lodash')
@@ -76,5 +77,17 @@ module.exports = {
     return branch
       ? branch.match(/^origin\/(.+)/)[1]
       : exec('git rev-parse --abbrev-ref HEAD').toString().trim()
+  },
+  setupRemote: function getRemote (info) {
+    let remote = `git@github.com:${info.repoSlug}`
+    if (info.gitUrl) remote = info.gitUrl
+
+    if (process.env.GH_TOKEN) {
+      if (remote.slice(0, 5) !== 'https') remote = `https://github.com/${info.repoSlug}`
+      const urlParsed = url.parse(remote)
+      urlParsed.auth = process.env.GH_TOKEN
+      remote = url.format(urlParsed)
+    }
+    exec(`git remote add gk-origin ${remote} || git remote set-url gk-origin ${remote}`)
   }
 }
